[ADD/#27] Data Class 구조
