--- app/src/main/java/com/example/eldercare/presentation/model/DummyModel.kt
@@ -1,5 +1,0 @@
-package com.example.eldercare.presentation.model
-
-data class DummyModel(
-    val test: String,
-)
