--- app/src/main/java/com/example/eldercare/presentation/ui/activity/AdditionalContactDiffCallback.kt
@@ -1,16 +1,0 @@
-package com.example.eldercare.presentation.ui.activity
-
-import com.example.eldercare.base.diffutil.BaseDiffCallback
-
-
-class AdditionalContactDiffCallback : BaseDiffCallback<AdditionalContactItem>(
-    itemTheSame = { oldItem, newItem ->
-        when {
-            oldItem is AdditionalContactItem.Contact && newItem is AdditionalContactItem.Contact ->
-                oldItem.id == newItem.id
-            else -> false
-        }
-    },
-    contentsTheSame = { oldItem, newItem -> oldItem == newItem }
-)
-
