--- app/src/main/java/com/example/eldercare/presentation/ui/activity/AdditionalContactItem.kt
@@ -1,6 +1,0 @@
-package com.example.eldercare.presentation.ui.activity
-
-sealed class AdditionalContactItem {
-    data class Contact(val id: Long, val phone: String) : AdditionalContactItem()
-}
-
