+++ app/src/main/java/com/example/eldercare/presentation/ui/custom/CustomEditText.kt
@@ -217,4 +217,9 @@ constructor(
         val imm = context.getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
         imm.showSoftInput(this, InputMethodManager.SHOW_IMPLICIT)
     }
+
+    fun requestFocusAndShowKeyboard() {
+        binding.etInputField.requestFocus()
+        binding.etInputField.showKeyboard()
+    }
 }
