+++ app/src/main/res/layout/view_custom_edit_text.xml
@@ -7,12 +7,10 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/cl_textField"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
+        android:layout_height="56dp"
         android:background="@drawable/shape_for_edit_text"
         android:paddingStart="20dp"
-        android:paddingTop="16dp"
         android:paddingEnd="20dp"
-        android:paddingBottom="16dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent">
@@ -20,7 +18,7 @@
         <EditText
             android:id="@+id/et_input_field"
             android:layout_width="0dp"
-            android:layout_height="28dp"
+            android:layout_height="0dp"
             android:autofillHints=""
             android:background="@null"
             android:imeOptions="actionDone"
